App: refactor CliApplication
    Move internal class CliParser out of CliApplication
    Move ConfigResolving logic to app/config/ConfigResolvingLoaderProvider
