+++ src/main/java/riid/app/ImageLoadingFacade.java
@@ -12,6 +12,7 @@ import java.util.Set;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import riid.app.config.AppConfig;
 import riid.app.error.AppError;
 import riid.app.error.AppException;
 import riid.core.fs.HostFilesystem;
